pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deltective"
version = "0.1.0"
description = "A detective for your Delta tables - inspect, analyze, and optimize"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "textual>=0.47.0",
    "deltalake>=0.15.0",
    "pyarrow>=14.0.0",
    "azure-identity>=1.12.0",
    "azure-storage-blob>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pandas>=2.0.0",
    "testcontainers[azurite]>=3.7.0",
]

[project.scripts]
deltective = "deltective.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py38"
